Fixup I8, headers.

+++ include/Core.h
@@ -13,7 +13,7 @@ namespace rat {
 	using U32 = uint32_t;
 	using U64 = uint64_t;
 
-	using I8 = int64_t;
+	using I8 = int8_t;
 	using I16 = int16_t;
 	using I32 = int32_t;
 	using I64 = int64_t;
